user 로 합침
