--- src/main/java/com/smartwave/tripsns/dao/IF_LoginDAO.java
@@ -1,8 +1,0 @@
-package com.smartwave.tripsns.dao;
-
-import com.smartwave.tripsns.vo.UserVO;
-
-public interface IF_LoginDAO {
-    public UserVO login(String id) throws Exception;
-
-}
