+++ src/main/java/com/smartwave/tripsns/dao/IF_UserDAO.java
@@ -2,8 +2,11 @@ package com.smartwave.tripsns.dao;
 
 import com.smartwave.tripsns.vo.UserVO;
 
-public interface IF_JoinDAO {
+public interface IF_UserDAO {
+    //회원가입
     public void userinsert(UserVO uservo) throws Exception;// 회원가입 정보 insert
     public int idchk(String id) throws Exception;
+    //로그인
+    public UserVO login(String id) throws Exception;
 
 }
