--- src/main/java/com/smartwave/tripsns/service/IF_LoginService.java
@@ -1,8 +1,0 @@
-package com.smartwave.tripsns.service;
-
-import com.smartwave.tripsns.vo.UserVO;
-
-public interface IF_LoginService {
-    public UserVO login(String id) throws Exception;
-
-}
