+++ src/main/java/com/smartwave/tripsns/service/IF_UserService.java
@@ -2,7 +2,10 @@ package com.smartwave.tripsns.service;
 
 import com.smartwave.tripsns.vo.UserVO;
 
-public interface IF_JoinService {
+public interface IF_UserService {
+    //회원가입
     public void userinsert(UserVO uservo) throws Exception;
     public int idchk(String id) throws Exception;
+    //로그인
+    public UserVO login(String id) throws Exception;
 }
