+++ src/main/resources/mapper/JoinMapper.xml
@@ -1,10 +1,17 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace= "com.smartwave.tripsns.dao.IF_JoinDAO">
+<mapper namespace= "com.smartwave.tripsns.dao.IF_UserDAO">
+    <!--회원가입-->
     <insert id="userinsert" parameterType="com.smartwave.tripsns.vo.UserVO">
         insert into testuser values (#{id},#{pw},#{name},#{birth},#{email})
     </insert>
     <select id="idchk" parameterType="String" resultType="int">
         select count(id) from testuser where id =#{id}
     </select>
+    <!--회원가입-->
+    <!--로그인-->
+    <select id="selectid" parameterType="String" resultType="com.smartwave.tripsns.vo.UserVO">
+        select * from testuser where id = #{id}
+    </select>
+    <!--로그인-->
 </mapper>
