--- src/main/resources/mapper/loginMapper.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace= "com.smartwave.tripsns.dao.IF_LoginDAO">
-    <select id="selectid" parameterType="String" resultType="com.smartwave.tripsns.vo.UserVO">
-        select * from testuser where id = #{id}
-    </select>
-
-</mapper>
