+++ src/main/resources/templates/loginForm.html
@@ -29,23 +29,10 @@
 <script >
  $(document).ready(function (){
      var errmsg = $('input[name=msge]').val();
-     console.log(errmsg);
-     // var errmsg = "${param.mesg}";
     if(errmsg !=""){
         alert(errmsg);
-        // errmsg=null;
     }else{
-        console.log(errmsg);
-
     }
-
-
-
-     // if(msga!=""){
-     //     alert(msga);
-     // }else{
-     //
-     // }
  });
